+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 * Describe initial release.
 
+## [0.1.1] - Add release date.
+* Fully fuctional
+* Can customize each tab with there own color.
+* complex animation
+* Can customize animation curves 
+* And a lot more ....